Move shared code to common module
