+++ cachito/workers/tasks/gitsubmodule.py
@@ -3,7 +3,7 @@ import logging
 
 import git
 
-from cachito.utils import PackagesData
+from cachito.common.packages_data import PackagesData
 from cachito.workers.pkg_managers.general import update_request_with_package
 from cachito.workers.paths import RequestBundleDir
 from cachito.workers.tasks.celery import app
