+++ cachito/workers/tasks/npm.py
@@ -4,8 +4,8 @@ import logging
 import os
 from typing import List
 
+from cachito.common.packages_data import PackagesData
 from cachito.errors import CachitoError
-from cachito.utils import PackagesData
 from cachito.workers import nexus
 from cachito.workers.config import get_worker_config, validate_npm_config
 from cachito.workers.paths import RequestBundleDir
