+++ cachito/workers/tasks/pip.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import logging
 import os
 
+from cachito.common.packages_data import PackagesData
 from cachito.errors import CachitoError
-from cachito.utils import PackagesData
 from cachito.workers import nexus
 from cachito.workers.config import get_worker_config, validate_pip_config
 from cachito.workers.paths import RequestBundleDir
