+++ tests/test_workers/test_pkg_managers/test_gomod.py
@@ -10,6 +10,7 @@ from unittest import mock
 import git
 import pytest
 
+from cachito.errors import CachitoError, ValidationError
 from cachito.workers.pkg_managers import gomod
 from cachito.workers.pkg_managers.gomod import (
     get_golang_version,
@@ -24,7 +25,6 @@ from cachito.workers.pkg_managers.gomod import (
     _set_full_local_dep_relpaths,
     _get_allowed_local_deps,
 )
-from cachito.errors import CachitoError, ValidationError
 from cachito.workers.paths import RequestBundleDir
 from tests.helper_utils import assert_directories_equal, write_file_tree
